% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukboundaries.R
\name{cache_dir}
\alias{cache_dir}
\title{Gets or sets the cache directory for data downloads}
\usage{
cache_dir(cachedir = NA)
}
\arguments{
\item{cachedir}{(optional) location of cache directory (if omitted just returns the current value)}
}
\value{
path to the current cache directory
}
\description{
Gets or sets the cache directory for data downloads
}
