% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{census11_codes}
\alias{census11_codes}
\title{2011 census hierarchical geographies}
\description{
2011 census hierarchical geographies
}
\examples{
\dontrun{
  # Get all OA codes within the City of London
  ladcode = "E09000001" # City of London
  oacodes = getsubgeographies(ladcode, "OA11")
  # Get the OA polygons and draw them
  sdf = getspatialdata(oacodes, "Boundaries", "GeneralisedClipped")
  plot(sdf$geometry)
}
}
