% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukborders-package.R
\name{duraz}
\alias{duraz}
\title{Download, unzip, read a zipped shapefile (tidy-up by default)}
\usage{
duraz(u, delete_zip = FALSE, rename_zip = TRUE)
}
\arguments{
\item{u}{character string pointing to a .zip file or an online url}

\item{delete_zip}{Should any downloaded .zip files be deleted (default is FALSE)?}

\item{rename_zip}{Should any downloaded .zip files be renames (defulat is TRUE)?}
}
\description{
Download, unzip, read a zipped shapefile (tidy-up by default)
}
