% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ed1981}
\alias{ed1981}
\title{Enumeration Districts 1981}
\description{
Saved for Leeds due to size constraints
}
\examples{
\dontrun{
u = "https://borders.ukdataservice.ac.uk/ukborders/easy_download/prebuilt/shape/England_ed_1981.zip"
ed1981 = duraz(u)
ed1981 = sf::st_transform(ed1981, 4326)
object.size(ed1981) / 1000000 # 143 mb
ed1981_simple = stplanr::mapshape(ed1981)
object.size(ed1981_simple) / 1e6 #  mb
leeds = lad2018[lad2018$lau118nm == "Leeds", ]
ed_cents = st_centroid(ed1981_simple)
ed_cents_lds = ed_cents[leeds, ]
ed1981 = ed1981[ed_cents_lds, ]
mapview::mapview(ed1981)
devtools::use_data(ed1981)
}
}
