% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukboundaries.R
\name{getspatialdata}
\alias{getspatialdata}
\title{for now area codes are assumed to be all the same}
\usage{
getspatialdata(area_codes, type, subtype)
}
\description{
for now area codes are assumed to be all the same
}
