% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{msoa2011_lds}
\alias{msoa2011_lds}
\title{MSOA boundaries in Leeds}
\description{
MSOA boundaries in Leeds
}
\examples{
\dontrun{
u = "https://borders.ukdataservice.ac.uk/ukborders/easy_download/prebuilt/shape/infuse_msoa_lyr_2011_clipped.zip"
msoa2011 = duraz(u)
msoa2011 = st_transform(msoa2011, 4326)
object.size(msoa2011) / 1e6 # 300+ mb
cents = st_centroid(msoa2011)
cents_lds = cents[leeds, ]
sel = msoa2011$geo_code \%in\% cents_lds$geo_code
msoa2011_lds = msoa2011[sel, ]
object.size(msoa2011_lds) / 1e6 # 4mb
mapview::mapview(msoa2011_lds)
plot(msoa2011_lds$geometry)
devtools::use_data(msoa2011_lds)
}
}
